A slightly faster default x-product filter function, and hopefully not as prone to sensitivity to the grammar factorization. Removed casts of ints -> shorts and vice versa. We'll allow 2^15 entries on either side of grammar rules. That should accomodate all the enumerable grammars we care about.

+++ java/src/edu/ohsu/cslu/parser/SparseMatrixVectorParser.java
@@ -184,7 +184,7 @@ public abstract class SparseMatrixVectorParser<G extends SparseMatrixGrammar, C 
             for (int i = 0; i < probabilities.length; i++) {
                 if (probabilities[i] != Float.NEGATIVE_INFINITY) {
                     final int leftChild = grammar.cartesianProductFunction().unpackLeftChild(i);
-                    final short rightChild = grammar.cartesianProductFunction().unpackRightChild(i);
+                    final int rightChild = grammar.cartesianProductFunction().unpackRightChild(i);
                     final int midpoint = midpoints[i];
                     final float probability = probabilities[i];
